+++ CHANGELOG.md
@@ -6,3 +6,4 @@ Yii 2 Base Models Change Log
 
 - Enh: Added `SubsidiaryTrait::hasSubsidiary()` to check whether the entity has the subsidiary or not.
 - Enh: Added `OperatorTrait` to record operator who operate the entity.
+- Enh: Added `createdAtToday` and `updatedAtToday` methods to specify the creation time or last updated time as today.